+++ test/batch_features_test.exs
@@ -14,6 +14,8 @@ defmodule EctoLibSql.BatchFeaturesTest do
 
     on_exit(fn ->
       File.rm(test_db)
+      File.rm(test_db <> "-shm")
+      File.rm(test_db <> "-wal")
     end)
 
     {:ok, database: test_db, opts: opts}