+++ test/connection_features_test.exs
@@ -12,6 +12,8 @@ defmodule EctoLibSql.ConnectionFeaturesTest do
 
     on_exit(fn ->
       File.rm(test_db)
+      File.rm(test_db <> "-shm")
+      File.rm(test_db <> "-wal")
     end)
 
     {:ok, database: test_db}