+++ test/ecto_sql_compatibility_test.exs
@@ -53,6 +53,8 @@ defmodule EctoLibSql.EctoSqlCompatibilityTest do
 
     on_exit(fn ->
       File.rm(@test_db)
+      File.rm(@test_db <> "-shm")
+      File.rm(@test_db <> "-wal")
     end)
 
     :ok